requirements.txt
requests
pandas
python-dotenv
click
matplotlib
seaborn
scikit-learn